refactor: move zip functions to FileOps
